# api/openapi.yaml
openapi: 3.0.3
info:
  title: Karvon Scraper API
  version: 1.0.0
  description: |
    REST API for Karvon: the scraper module (Google Maps lead scraping with website
    email crawling) and the campaign module (outreach campaigns, contacts and their
    lifecycle, reusable email content, consent and the two delivery providers). Live
    job progress is delivered over Server-Sent Events.

    All routes except `/healthz`, `/openapi.json` and `/webhooks/*` require
    `Authorization: Bearer <KARVON_API_KEY>`. The `/webhooks/*` routes carry no API
    key: they are called by Instantly and Mailchimp, and are authenticated instead by
    the per-registration token in their path together with the secret that provider
    signs its deliveries with.
servers:
  - url: /api/v1
    description: Default mount point
security:
  - bearerAuth: []
...
paths:
...
  /verification/stats:
    get:
      tags: [verification]
      operationId: getVerificationStats
      summary: Tag distribution, pass coverage, credits and the current cost to verify
      responses:
        '200':
          description: Verification statistics
          content:
            application/json:
              schema: { $ref: '#/components/schemas/VerificationStats' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /verification/settings:
    get:
      tags: [verification]
      operationId: getVerificationSettings
      summary: Provider weights, toggles, the paid band and live provider health
      responses:
        '200':
          description: The current verification settings
          content:
            application/json:
              schema: { $ref: '#/components/schemas/VerificationSettingsView' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    put:
      tags: [verification]
      operationId: updateVerificationSettings
      summary: Replace the verification settings
      description: |
        The weights of every weighted provider must total 100%, whether or not each
        provider is enabled; a disabled or unreachable provider has its weight
        redistributed at scoring time rather than requiring the operator to
        re-balance. `paid_min_score` must not exceed `paid_threshold`.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/VerificationSettingsUpdate' }
      responses:
        '200':
          description: The saved settings
          content:
            application/json:
              schema: { $ref: '#/components/schemas/VerificationSettingsView' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
  /verification/emails:
    get:
      tags: [verification]
      operationId: listVerifications
      summary: Every address that has been through, or is queued for, verification
      parameters:
        - name: tag
          in: query
          description: Repeatable tag filter
          schema:
            type: array
            items: { $ref: '#/components/schemas/VerificationTag' }
          style: form
          explode: true
        - name: min_score
          in: query
          schema: { type: integer, minimum: 0, maximum: 100 }
        - name: max_score
          in: query
          schema: { type: integer, minimum: 0, maximum: 100 }
        - name: pass2_status
          in: query
          schema: { $ref: '#/components/schemas/Pass2Status' }
        - name: needs_third_party
          in: query
          description: |
            Only addresses that passed the local gate, fall inside the paid band and
            have never been sent to a third-party verifier. An address is sent to one
            at most once, ever, so this set only ever shrinks.
          schema: { type: boolean }
        - name: has_typo
          in: query
          schema: { type: boolean }
        - name: q
          in: query
          description: Trigram search over the address
          schema: { type: string, maxLength: 200 }
        - name: business_id
          in: query
          schema: { type: string, format: uuid }
        - name: job_id
          in: query
          schema: { type: string, format: uuid }
        - name: include_suppressed
          in: query
          description: Defaults to false, which hides addresses held only by suppressed businesses
          schema: { type: boolean }
        - name: sort
          in: query
          schema:
            type: string
            default: final_score:desc
            enum: [final_score:desc, final_score:asc, email:asc, email:desc,
                   pass1_verified_at:desc, pass1_verified_at:asc,
                   pass2_verified_at:desc, pass2_verified_at:asc,
                   third_party_sent_at:desc, third_party_sent_at:asc,
                   updated_at:desc, updated_at:asc]
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PerPage'
      responses:
        '200':
          description: Paginated verifications
          content:
            application/json:
              schema: { $ref: '#/components/schemas/EmailVerificationList' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
  /verification/emails/{id}:
    parameters:
      - $ref: '#/components/parameters/IdPath'
    get:
      tags: [verification]
      operationId: getVerification
      summary: One address with its full Pass 1 breakdown and raw provider payload
      responses:
        '200':
          description: Verification
          content:
            application/json:
              schema: { $ref: '#/components/schemas/EmailVerificationDetail' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
  /verification/emails/{id}/self:
    parameters:
      - $ref: '#/components/parameters/IdPath'
    post:
      tags: [verification]
      operationId: verifyEmailSelf
      summary: Queue the free local pass for one address
      responses:
        '202':
          description: The run that will score this address
          content:
            application/json:
              schema: { $ref: '#/components/schemas/VerificationRun' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
  /verification/emails/{id}/third-party:
    parameters:
      - $ref: '#/components/parameters/IdPath'
    post:
      tags: [verification]
      operationId: verifyEmailThirdParty
      summary: Queue the paid provider check for one address
      description: |
        Refused with 409 when the address has already been sent to a third-party
        verifier, has not completed the local pass, scored below the gate, or scored
        at or above the confidence threshold. The first of those is permanent: an
        address is sent to a third party at most once, ever, and no setting, score or
        retry lifts that. In every case nothing would be billed, so the request is
        rejected rather than silently skipped.
      responses:
        '202':
          description: The run that will verify this address
          content:
            application/json:
              schema: { $ref: '#/components/schemas/VerificationRun' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
  /verification/emails/{id}/apply-typo:
    parameters:
      - $ref: '#/components/parameters/IdPath'
    post:
      tags: [verification]
      operationId: applyTypoSuggestion
      summary: Rewrite every business holding the misspelled address to the suggested one
      description: |
        A business that already holds the corrected address keeps it and loses the
        misspelled row. Primary addresses are re-picked afterwards and the corrected
        address is queued for local verification. The response is the corrected
        address's verification row.
      responses:
        '200':
          description: The corrected address
          content:
            application/json:
              schema: { $ref: '#/components/schemas/EmailVerificationDetail' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /verification/runs/estimate:
    post:
      tags: [verification]
      operationId: estimateVerificationRun
      summary: How many addresses a run would process, and what it would cost
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/VerificationRunEstimateRequest' }
      responses:
        '200':
          description: Estimate
          content:
            application/json:
              schema: { $ref: '#/components/schemas/VerificationRunEstimate' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
  /verification/runs:
    get:
      tags: [verification]
      operationId: listVerificationRuns
      summary: Run history
      parameters:
        - name: pass
          in: query
          schema: { $ref: '#/components/schemas/VerificationPass' }
        - name: status
          in: query
          schema: { $ref: '#/components/schemas/JobStatus' }
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PerPage'
      responses:
        '200':
          description: Paginated runs
          content:
            application/json:
              schema: { $ref: '#/components/schemas/VerificationRunList' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
    post:
      tags: [verification]
      operationId: createVerificationRun
      summary: Start a bulk verification run
      description: |
        A `third_party` run must carry `max_cost_cents`: the server re-estimates and
        refuses with 409 if the cost has grown past what was approved, so a run can
        never spend more than the figure the operator confirmed.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/VerificationRunCreate' }
      responses:
        '201':
          description: Created run
          content:
            application/json:
              schema: { $ref: '#/components/schemas/VerificationRun' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
  /verification/runs/{id}:
    parameters:
      - $ref: '#/components/parameters/IdPath'
    get:
      tags: [verification]
      operationId: getVerificationRun
      summary: Run progress, polled while a run is active
      responses:
        '200':
          description: Run
          content:
            application/json:
              schema: { $ref: '#/components/schemas/VerificationRun' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
  /verification/runs/{id}/cancel:
    parameters:
      - $ref: '#/components/parameters/IdPath'
    post:
      tags: [verification]
      operationId: cancelVerificationRun
      summary: Cooperatively cancel a queued or running verification run
      responses:
        '200':
          description: Cancelled run
          content:
            application/json:
              schema: { $ref: '#/components/schemas/VerificationRun' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: "Static API key, `Authorization: Bearer <KARVON_API_KEY>`"
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
    Page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    PerPage:
      name: per_page
      in: query
      schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
...
  responses:
    BadRequest:
      description: Malformed request
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Unauthorized:
      description: Missing or invalid API key
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: Resource does not exist
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Conflict:
      description: Action not allowed in the resource's current state
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    ValidationFailed:
      description: Request body or query failed validation
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string, example: validation_failed }
            message: { type: string }
            details:
              type: array
              items: { $ref: '#/components/schemas/FieldError' }
      additionalProperties: false
    FieldError:
      type: object
      required: [field, message]
      properties:
        field: { type: string }
        message: { type: string }
    PageMeta:
      type: object
      required: [page, per_page, total]
      properties:
        page: { type: integer }
        per_page: { type: integer }
        total: { type: integer, format: int64 }
...
    JobStatus:
      type: string
      enum: [queued, running, done, failed, cancelled]
...
    VerificationTag:
      type: string
      description: |
        The colour band a final score falls into. The UI must never convey a tag by
        colour alone; `tag_label` carries the text that goes with it.
      enum: [green, light_green, yellow, orange, red]
    Pass2Status:
      type: string
      description: |
        The third-party verdict. Only `deliverable`, `risky` and `undeliverable` are
        conclusive and billed. `unknown` and `error` leave the free score in place.
        Whether the address may be tried again is not read from this field but from
        `third_party_sent_at`: `unknown` came from the provider and is final, while
        `error` from a call that never reached it leaves the address unsent.
      enum: [deliverable, risky, unknown, undeliverable, error]
    VerificationPass:
      type: string
      description: "`self` is the free local pass; `third_party` is the paid one."
      enum: [self, third_party]
    VerificationCheckStatus:
      type: string
      enum: [pass, fail, skip]
    VerificationCheck:
      type: object
      description: One line of the Pass 1 breakdown. The points always sum to pass1_score.
      required: [key, label, status, points, max]
      properties:
        key: { type: string, example: mx }
        label: { type: string, example: MX records present }
        status: { $ref: '#/components/schemas/VerificationCheckStatus' }
        points: { type: integer, description: Points this check actually contributed }
        max: { type: integer, description: Points it could have contributed }
        detail: { type: string }
    VerificationBusiness:
      type: object
      required: [id, name]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
    EmailVerification:
      type: object
      required: [id, email, domain, pass1_score, free_score, pass2_credits, final_score,
                 verification_tag, tag_label, business_count, updated_at]
      properties:
        id: { type: string, format: uuid }
        email: { type: string }
        domain: { type: string }
        pass1_score: { type: integer, minimum: 0, maximum: 85 }
        pass1_hard_fail:
          type: string
          nullable: true
          description: Key of the check that zeroed the score, null when none did
        pass1_verified_at: { type: string, format: date-time, nullable: true }
        free_score:
          type: integer
          minimum: 0
          maximum: 89
          description: |
            The weighted combination of every free provider. Capped below 90 on
            purpose: only the paid provider can tag an address Verified.
        free_scored_at: { type: string, format: date-time, nullable: true }
        pass2_score: { type: integer, nullable: true }
        pass2_status:
          allOf:
            - $ref: '#/components/schemas/Pass2Status'
          nullable: true
        pass2_verified_at: { type: string, format: date-time, nullable: true }
        third_party_sent_at:
          type: string
          format: date-time
          nullable: true
          description: |
            When this address was handed to a third-party verifier. Non-null means it
            has had its one and only send: it can never be sent to a third party
            again, whatever the provider answered or failed to answer. Null means it
            has never left our infrastructure.
        pass2_credits: { type: integer }
        final_score: { type: integer, minimum: 0, maximum: 100 }
        verification_tag: { $ref: '#/components/schemas/VerificationTag' }
        tag_label: { type: string, example: Likely valid }
        typo_suggestion:
          type: string
          nullable: true
          description: The corrected address, offered to the operator rather than applied
        last_error: { type: string, nullable: true }
        business_count: { type: integer }
        updated_at: { type: string, format: date-time }
    EmailVerificationDetail:
      allOf:
        - $ref: '#/components/schemas/EmailVerification'
        - type: object
          properties:
            pass1_checks:
              type: array
              items: { $ref: '#/components/schemas/VerificationCheck' }
            provider_results:
              type: array
              description: |
                How each free provider contributed to free_score. Every weighted
                provider always appears, including ones that were switched off or
                could not be reached, so the breakdown is always complete.
              items: { $ref: '#/components/schemas/VerificationProviderResult' }
            pass2_raw:
              type: object
              additionalProperties: true
              nullable: true
            businesses:
              type: array
              items: { $ref: '#/components/schemas/VerificationBusiness' }
    EmailVerificationList:
      type: object
      required: [data, meta]
      properties:
        data:
          type: array
          items: { $ref: '#/components/schemas/EmailVerification' }
        meta: { $ref: '#/components/schemas/PageMeta' }
...
    VerificationProviderHealth:
      type: object
      required: [provider, label, stage, weighted, toggleable, description, enabled, weight]
      properties:
        provider: { $ref: '#/components/schemas/VerificationProviderKey' }
        label: { type: string }
        stage:
          type: string
          enum: [free, paid]
          description: '`free` providers are weighted together; `paid` runs afterwards and costs credits'
        weighted: { type: boolean, description: Whether it takes part in the weighted score }
        toggleable: { type: boolean, description: Whether it can be switched off }
        description: { type: string }
        enabled: { type: boolean }
        weight: { type: integer, description: The configured percentage; 0 for the paid provider }
        healthy:
          type: boolean
          nullable: true
          description: |
            null when readiness cannot be determined, which is the case for providers
            that need no probing. Render that as "not applicable", never as healthy.
        error: { type: string, description: Why the readiness probe failed }
    VerificationSettings:
      type: object
      required: [weights, enabled, paid_enabled, paid_threshold, paid_min_score]
      properties:
        weights:
          type: object
          additionalProperties: { type: integer, minimum: 0, maximum: 100 }
          description: Percentage per weighted provider, keyed by provider. Must total 100.
          example: { existing: 30, mailchecker: 20, reacher: 50 }
        enabled:
          type: object
          additionalProperties: { type: boolean }
          description: Which toggleable providers are switched on, keyed by provider
          example: { mailchecker: true, reacher: false, paid: true }
        paid_enabled:
          type: boolean
          description: Switches the paid stage off entirely
        paid_threshold:
          type: integer
          minimum: 0
          maximum: 100
          description: |
            The ceiling, exclusive. An address whose free score reaches it is already
            confident enough and is never billed.
        paid_min_score:
          type: integer
          minimum: 0
          maximum: 100
          description: |
            The floor, inclusive. An address below it is never billed, because a paid
            provider would only confirm what the free checks already established.
    VerificationSettingsUpdate:
      allOf:
        - $ref: '#/components/schemas/VerificationSettings'
      description: The full settings object; every field is replaced.
    VerificationSettingsView:
      type: object
      required: [settings, providers, free_max_score, weight_total]
      properties:
        settings: { $ref: '#/components/schemas/VerificationSettings' }
        providers:
          type: array
          description: The provider catalogue with current settings and live readiness
          items: { $ref: '#/components/schemas/VerificationProviderHealth' }
        free_max_score:
          type: integer
          description: |
            The ceiling on the weighted free score. Below the Verified band on
            purpose, so the UI can explain why three perfect free scores do not
            produce 100.
        weight_total:
          type: integer
          description: What the weights must add up to, so the client can validate before submitting
        updated_at: { type: string, format: date-time, nullable: true }
...
    VerificationRunEstimateRequest:
      type: object
      required: [pass]
      properties:
        pass: { $ref: '#/components/schemas/VerificationPass' }
        filter: { $ref: '#/components/schemas/VerificationRunFilter' }
      additionalProperties: false
    VerificationRunEstimate:
      type: object
      required: [emails, needs_self, cached, credits_needed, est_cost_cents]
      properties:
        emails:
          type: integer
          format: int64
          description: How many addresses the run would actually process
        needs_self:
          type: integer
          format: int64
          description: Matching addresses excluded because they have not been through the free checks
        cached:
          type: integer
          format: int64
          description: |
            Matching addresses skipped because they have already been sent to a
            third party once. They are skipped permanently and are never billed.
        credits_needed: { type: integer, format: int64 }
        cost_per_1k_cents: { type: integer }
        est_cost_cents: { type: integer, format: int64 }
        balance_credits:
          type: integer
          format: int64
          nullable: true
          description: The provider's reported balance, null when it cannot be reached
    VerificationRunCreate:
      type: object
      required: [pass]
      properties:
        pass: { $ref: '#/components/schemas/VerificationPass' }
        filter: { $ref: '#/components/schemas/VerificationRunFilter' }
        max_cost_cents:
          type: integer
          format: int64
          minimum: 0
          description: |
            Required for a third_party run. Send the estimate you displayed; the
            server re-estimates and refuses if the real cost exceeds it.
      additionalProperties: false
    VerificationRun:
      type: object
      required: [id, pass, status, filter, total, done, failed, skipped, credits_used,
                 est_cost_cents, created_at]
      properties:
        id: { type: string, format: uuid }
        pass: { $ref: '#/components/schemas/VerificationPass' }
        status: { $ref: '#/components/schemas/JobStatus' }
        filter: { $ref: '#/components/schemas/VerificationRunFilter' }
        total: { type: integer }
        done: { type: integer }
        failed: { type: integer }
        skipped:
          type: integer
          description: |
            Addresses that had already been sent to a third party, were below the
            gate, or no longer exist
        credits_used: { type: integer }
        est_cost_cents: { type: integer, format: int64 }
        source_id: { type: string, format: uuid, nullable: true }
        error: { type: string, nullable: true }
        created_at: { type: string, format: date-time }
        started_at: { type: string, format: date-time, nullable: true }
        finished_at: { type: string, format: date-time, nullable: true }
    VerificationRunList:
      type: object
      required: [data, meta]
      properties:
        data:
          type: array
          items: { $ref: '#/components/schemas/VerificationRun' }
        meta: { $ref: '#/components/schemas/PageMeta' }
    VerificationTagCounts:
      type: object
      required: [green, light_green, yellow, orange, red]
      properties:
        green: { type: integer, format: int64 }
        light_green: { type: integer, format: int64 }
        yellow: { type: integer, format: int64 }
        orange: { type: integer, format: int64 }
        red: { type: integer, format: int64 }
    VerificationStats:
      type: object
      required: [total, by_tag, self_verified, third_party_verified, needs_self,
                 qualifying_for_third_party, qualifying_est_cost_cents,
                 credits_used_total, credits_used_30d, active_runs]
      properties:
        total: { type: integer, format: int64 }
        by_tag: { $ref: '#/components/schemas/VerificationTagCounts' }
        self_verified: { type: integer, format: int64 }
        third_party_verified: { type: integer, format: int64 }
        needs_self:
          type: integer
          format: int64
          description: Addresses on the master list that have never completed Pass 1
        qualifying_for_third_party:
          type: integer
          format: int64
          description: Addresses that would be sent, and billed for, by a paid run today
        qualifying_est_cost_cents: { type: integer, format: int64 }
        credits_used_total: { type: integer, format: int64 }
        credits_used_30d: { type: integer, format: int64 }
        balance_credits: { type: integer, format: int64, nullable: true }
        active_runs: { type: integer, format: int64 }
        last_self_run_at: { type: string, format: date-time, nullable: true }
        last_third_party_run_at: { type: string, format: date-time, nullable: true }
...